Move GameImageView out from SaveDataScreen
