+++ UI/MiscViews.h
@@ -3,6 +3,7 @@
 #include "Common/Common.h"
 #include "Common/UI/View.h"
 #include "UI/ViewGroup.h"
+#include "UI/GameInfoCache.h"
 
 // Compound view, showing a text with an icon.
 class TextWithImage : public UI::LinearLayout {
@@ -56,4 +57,22 @@ public:
 
 private:
 	Path gamePath_;
+};
+
+class GameImageView : public UI::InertView {
+public:
+	GameImageView(const Path &gamePath, GameInfoFlags image, float scale, UI::LayoutParams *layoutParams = 0)
+		: InertView(layoutParams), image_(image), gamePath_(gamePath), scale_(scale) {
+	}
+
+	void GetContentDimensions(const UIContext &dc, float &w, float &h) const override;
+	void Draw(UIContext &dc) override;
+	std::string DescribeText() const override { return ""; }
+
+private:
+	Path gamePath_;
+	GameInfoFlags image_;
+	float scale_ = 1.0f;
+	int textureWidth_ = 0;
+	int textureHeight_ = 0;
 };
